==> 4. Data Analysis.ipynb <==
from scraperDB import DBManager
import pandas as pd
import matplotlib.pyplot as plt
import pandas as pd
import numpy as np
import os

def load_dfs(parent_folder):
    dfs = {}
    for location in os.listdir(parent_folder):
        location_path = os.path.join(parent_folder, location)
        table_name = os.path.basename(location_path)
        if os.path.isdir(location_path):
            file_path = os.path.join(location_path, 'data_clean.csv')
            if os.path.exists(file_path):
                df = pd.read_csv(file_path)
                dfs[table_name] = df
                print(f'DataFrame from {location_path} loaded')
    return dfs

data = {}
data['rent'] = load_dfs('rent')
data['sale'] = load_dfs('sale')

dbm = DBManager.DBM()

def generate_dbs(_dict, suffix):
    for k, v in _dict.items():
        db = f'{k}_{suffix}'
        dbm.create_db(db)
        dbm.connect(db, force=True)
        dbm.create_table('original', v, force=True)
        dbm.disconnect(force=True)

generate_dbs(data['rent'], 'rent')
generate_dbs(data['sale'], 'sale')
